Reviewed and fixed FP register usage throughout
The argument and return value registers were not being utilized
correctly in several places. I also removed the legacy handling of
floating point registers in the simulator, inherited from MIPS, where
the double precision registers are assumed to take two single-precision
registers.

There is one remaining FP related cctest that appears to be a separate
issue, so I have opened #111 for this test.

Also added the `--fp` flag to the trace analyzer tool. When passed, the
floating point argument and return value registers are printed, along
with the scalar registers.

Fixes #98