RISC-V: Add instruction cache flushing
I know the ISA manual says fence.i in it, and that the older versions say it's
in I, but it's not and you can't use it from userspace in Linux (at least, if
you want your programs to work).  It's currently implemented as a system call,
but as there may eventually be something faster you should use the GCC builtin
instead.

Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>