Fix mul and mulh to match the ISA manual
mulh on RISC-V doesn't work like mulh on MIPS: the rv64 mulh produces
the top 64 bits of a 128-bit multiplication, as opposed to the top 32
bits of a 64-bit multiplication.  The simulator followed the old
behavior, which is incorrect on RISC-V.  This patch fixes both the
simulator's implementation of mulh and the compiler's one use of mulh to
match the ISA's definition.  I haven't run it on the hardware, but it
should improve things.

Additionally, I fixed the simulator's mul implementation to perform a
64-bit multiplication, which matches the ISA manual.  I've kept these
two changes in the same patch because the mul changes is necessary for
the new Mulh to function, which is necessary to handle the mulh change.

It doesn't clean up the larger issues in the port related to mulh, as
now we're performing all 64-bit multiplications twice and throwing away
half of the result each time.  I'm trying to split up the bug fixes from
the cleanups.

Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>