[CI] Change the output style to verbose
This patch changes the output style of run-tests.py from the
default 'mono' style to 'verbose' for more useful CI log.

The problem of the mono style is that we could not figure out
easily which testcase failed and what the error message was.
And redirecting the output of run-tests.py could also capture
lots of control & invisable ASCII chars, which is not useful.