Fix bug with Wasm frames
The bug is due to setting the incorrect `WasmCompileLazyFrameConstants::kWasmInstanceOffset` value
in `frame-constants-riscv64.h` that leads to loading the wasm instance from the incorrect offset
off the frame.

It seems that the problem should have existed in the `riscv-port-dev` branch and in MIPS64 backend
as well, but not triggered either because wasm lazy compilation is not triggered or because there
was no debug check in `riscv-port-dev` branch.

Fixes #135