Ensure that a failure is reported by CI
Added `set -e` so that if any command fails, the script will fail and it
will be reported appropriately by the CI server. Also cleaned up the
script a bit and merged all of the tests into one command so that the
results are reported all together at the end.

Fixes #256